Add backspace button and refactor display code
There is now a backspace button that removes the last number or operator
entered. The display functionality was refactored to be a function that
other functions will call so that anything display related happens in
the updateDisplay function. Also pulled decimal functionality out of
number event listener and made it its own listener.